# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Прогноз отказов оборудования коллекторов",
    "description": "REST API сервиса прогноза отказов датчиков коллекторов на 24 часа (ЛЦТ-2026).\n\n**Вход:** нажмите *Authorize* и введите логин и пароль (в демо-режиме — демо-пользователя из README), поля client_id/secret оставьте пустыми.\n\n**Машина времени:** все запросы на чтение принимают `at`. Демо-данные покрывают 01.07.2026–31.08.2026, например `at=2026-08-01T12:00:00`.\n\n**Симуляция потока** (`/api/sim/*`): проигрывание суток по заранее рассчитанным прогнозам с ускорением; пока она идёт, запросы без `at` работают на модельное время. Онлайн-пересчёт моделью не выполняется.\n\nВремя — московское, без часового пояса.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "Журнал событий"
        ],
        "summary": "Журнал тревожных событий и смен статуса с контекстом (плановая проверка, групповое отключение, сбой)",
        "description": "События оперативного контура (`events_recent`): по умолчанию — тревожные сообщения о состоянии за сутки до `at`. Контекст — прозрачные правила, не ML: «Вероятная плановая проверка» (газ в будни 9–14), «Групповое отключение — вероятно, плановые работы» (≥ 80 % и не меньше 5 однотипных датчиков объекта отключились в ±12 ч), «Сбой значения» (служебный код, дата вместо показания, отрицательный газ). Только датчики области пользователя. `tag` фильтрует страницу результата (правила считаются на странице).",
        "operationId": "events_api_events_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Начало периода; по умолчанию сутки до конца периода",
              "title": "Date From"
            },
            "description": "Начало периода; по умолчанию сутки до конца периода"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Конец периода; по умолчанию at (не позже последнего события)",
              "title": "Date To"
            },
            "description": "Конец периода; по умолчанию at (не позже последнего события)"
          },
          {
            "name": "object_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Объект; включает вложенные",
              "title": "Object Id"
            },
            "description": "Объект; включает вложенные"
          },
          {
            "name": "channel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Один датчик (блок событий на карточке)",
              "title": "Channel Id"
            },
            "description": "Один датчик (блок событий на карточке)"
          },
          {
            "name": "system_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "System Type"
            }
          },
          {
            "name": "sensor_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sensor Type"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поиск по названию, тегу или id датчика",
              "title": "Q"
            },
            "description": "Поиск по названию, тегу или id датчика"
          },
          {
            "name": "only_alarms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Только события с флагом «тревожное»",
              "default": true,
              "title": "Only Alarms"
            },
            "description": "Только события с флагом «тревожное»"
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "status",
                "values"
              ],
              "type": "string",
              "description": "status — сообщения о состоянии (смены статуса), values — только показания, all — всё",
              "default": "status",
              "title": "Kind"
            },
            "description": "status — сообщения о состоянии (смены статуса), values — только показания, all — всё"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "planned_check",
                    "group_off",
                    "value_failure"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Только события с этим контекстом (проверяется на странице результата)",
              "title": "Tag"
            },
            "description": "Только события с этим контекстом (проверяется на странице результата)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 50,
              "title": "Limit"
            },
            "description": "Размер страницы"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Сдвиг от начала",
              "default": 0,
              "title": "Offset"
            },
            "description": "Сдвиг от начала"
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/export/events": {
      "get": {
        "tags": [
          "Журнал событий"
        ],
        "summary": "Выгрузка журнала событий в XLSX (те же фильтры, что у /api/events)",
        "operationId": "export_events_api_export_events_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Начало периода; по умолчанию сутки до конца периода",
              "title": "Date From"
            },
            "description": "Начало периода; по умолчанию сутки до конца периода"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Конец периода; по умолчанию at (не позже последнего события)",
              "title": "Date To"
            },
            "description": "Конец периода; по умолчанию at (не позже последнего события)"
          },
          {
            "name": "object_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Объект; включает вложенные",
              "title": "Object Id"
            },
            "description": "Объект; включает вложенные"
          },
          {
            "name": "channel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Один датчик (блок событий на карточке)",
              "title": "Channel Id"
            },
            "description": "Один датчик (блок событий на карточке)"
          },
          {
            "name": "system_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "System Type"
            }
          },
          {
            "name": "sensor_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sensor Type"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поиск по названию, тегу или id датчика",
              "title": "Q"
            },
            "description": "Поиск по названию, тегу или id датчика"
          },
          {
            "name": "only_alarms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Только события с флагом «тревожное»",
              "default": true,
              "title": "Only Alarms"
            },
            "description": "Только события с флагом «тревожное»"
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "status",
                "values"
              ],
              "type": "string",
              "description": "status — сообщения о состоянии (смены статуса), values — только показания, all — всё",
              "default": "status",
              "title": "Kind"
            },
            "description": "status — сообщения о состоянии (смены статуса), values — только показания, all — всё"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "planned_check",
                    "group_off",
                    "value_failure"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Только события с этим контекстом (проверяется на странице результата)",
              "title": "Tag"
            },
            "description": "Только события с этим контекстом (проверяется на странице результата)"
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "responses": {
          "200": {
            "description": "Файл XLSX, до 50 000 строк",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "examples": [
              334609
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Дым ПК 1101+2"
            ]
          },
          "sensor_type": {
            "type": "string",
            "title": "Sensor Type",
            "examples": [
              "Датчик дыма"
            ]
          },
          "system_type": {
            "type": "string",
            "title": "System Type",
            "examples": [
              "Пожарная охрана"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "sensor_type",
          "system_type"
        ],
        "title": "ChannelRef"
      },
...
      "EventItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelRef"
          },
          "object": {
            "$ref": "#/components/schemas/ObjectRef"
          },
          "picket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Picket",
            "description": "Пикет из названия датчика (docs/GEO_SCHEME.md)",
            "examples": [
              "ПК447+5"
            ]
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "Значение как в журнале заказчика"
          },
          "value_num": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Num"
          },
          "value_kind": {
            "type": "string",
            "enum": [
              "status",
              "number",
              "service",
              "date"
            ],
            "title": "Value Kind"
          },
          "value_kind_label": {
            "type": "string",
            "title": "Value Kind Label"
          },
          "is_alarm": {
            "type": "boolean",
            "title": "Is Alarm",
            "description": "Флаг «тревожное» из журнала заказчика"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/EventTag"
            },
            "type": "array",
            "title": "Tags",
            "description": "Контекст по прозрачным правилам (не ML), app/event_context.py"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ts",
          "channel",
          "object",
          "value",
          "value_kind",
          "value_kind_label",
          "is_alarm",
          "tags"
        ],
        "title": "EventItem"
      },
      "EventList": {
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "date_from": {
            "type": "string",
            "format": "date-time",
            "title": "Date From"
          },
          "date_to": {
            "type": "string",
            "format": "date-time",
            "title": "Date To"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "at",
          "date_from",
          "date_to",
          "total",
          "items"
        ],
        "title": "EventList"
      },
...
      "EventTag": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "planned_check",
              "group_off",
              "value_failure"
            ],
            "title": "Code"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "examples": [
              "Вероятная плановая проверка"
            ]
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Почему правило сработало — факты, языком диспетчера"
          }
        },
        "type": "object",
        "required": [
          "code",
          "label",
          "reason"
        ],
        "title": "EventTag"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ObjectRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "examples": [
              20
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "объект Фита"
            ]
          },
          "path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Path",
            "description": "Путь от района до объекта включительно",
            "examples": [
              [
                "Район по эксплуатации",
                "объект Бета",
                "объект Фита"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path"
        ],
        "title": "ObjectRef"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}